requirements.txt
pylint
pytest
click
black
pytest-cov
pandas
nbval
click
flask
requests
pandas-gbq
sklearn
tensorflow
fsspec
gcsfs
google-cloud-storage
google-cloud-automl
google-cloud-bigquery
datalab
#google.datalab.storage
#google.datalab.bigquery